requirements_mtest.txt
mitmproxy==8.0.0
requests
pymongo
psycopg2-binary
bson
pandas
